fixup 'fec: adapt to lx_kit/lx_emul update'

+++ repos/dde_linux/src/drivers/nic/fec/lx_emul.cc
@@ -743,7 +743,7 @@ long __wait_completion(struct completion *work, unsigned long timeout)
 
 	if (timeout) {
 		timer_setup(&t, _completion_timeout, 0u);
-		t.data = Lx::scheduler().current();
+		t.data = (unsigned long) Lx::scheduler().current();
 		mod_timer(&t, j);
 	}
